pyproject.toml
[tool.poetry]
name = "py_engineering_chat"
version = "0.1.0"
description = "A research and chat application with web crawling capabilities"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
pymilvus = "2.2.8"
python-dotenv = "^1.0.0"
langchain-chroma = "^0.1.3"
scrapy = "^2.11.2"

[tool.poetry.dev-dependencies]
# Add any development dependencies here, if needed

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
